Conflicts:
	portal/slicerequestview.py
	portal/usersview.py

+++ portal/actions.py
@@ -110,21 +110,16 @@ def manifold_add_account(request, account_params):
     result, = results
     return result['user_id']
 
-def manifold_update_account(request,account_params):
+def manifold_update_account(request,user_id,account_params):
     # account_params: config
-    query = Query.update('local:account').filter_by('platform', '==', 'myslice').set(account_params).select('user_id')
-    results = execute_query(request,query)
-    # NOTE: results remains empty and goes to Exception. However, it updates the manifold DB.
-    # That's why I commented the exception part. -- Yasin 
-    #if not results:
-    #    raise Exception, "Failed updating manifold account: config %s" % account_params['config']
-    #result, = results
+    query = Query.update('local:account').filter_by('platform', '==', 'myslice').filter_by('user_id', '==', user_id).set(account_params).select('user_id')
+    results = execute_admin_query(request,query)
     return results
 
 #explicitly mention the platform_id
-def manifold_delete_account(request, platform_id, account_params):
-    query = Query.delete('local:account').filter_by('platform_id', '==', platform_id).set(account_params).select('user_id')
-    results = execute_query(request,query)
+def manifold_delete_account(request, platform_id, user_id, account_params):
+    query = Query.delete('local:account').filter_by('platform_id', '==', platform_id).filter_by('user_id', '==', user_id).set(account_params).select('user_id')
+    results = execute_admin_query(request,query)
     return results
